Clase9/src/main/java/org/example/Calculadora.java
package org.example;

public class Calculadora {
    // Podrían ser static
    public double sumar(double a, double b) {
        return a + b;
    }

    public double restar(double a, double b) {
        return a - b;
    }

    public double multiplicar(double a, double b) {
        return a * b;
    }

    public double dividir(double a, double b) {
        if (b == 0) {
            throw new ArithmeticException("No se puede dividir por cero");
        } else {
            return a / b;
        }
    }
}
